# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Task Force API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
} as const;

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  TODO: 'TODO',
  IN_PROGRESS: 'IN_PROGRESS',
  DONE: 'DONE',
} as const;

export interface Task {
  id: number;
  userId: number;
  /** @nullable */
  classificationId?: number | null;
  /** @nullable */
  classificationName?: string | null;
  /** @nullable */
  groupId?: number | null;
  /** @nullable */
  assignedUserId?: number | null;
  title: string;
  /** @nullable */
  description?: string | null;
  priority: TaskPriority;
  status: TaskStatus;
  /** @nullable */
  deadline?: string | null;
  isOverdue: boolean;
  subtaskCount: number;
  completedSubtaskCount: number;
  emailNotificationSent?: boolean;
  createdAt: string;
  updatedAt: string;
}

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
} as const;

export type TaskInputStatus = typeof TaskInputStatus[keyof typeof TaskInputStatus];


export const TaskInputStatus = {
  TODO: 'TODO',
  IN_PROGRESS: 'IN_PROGRESS',
  DONE: 'DONE',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  priority: TaskInputPriority;
  status: TaskInputStatus;
  deadline?: string;
  classificationId?: number;
  groupId?: number;
}

export type TaskUpdatePriority = typeof TaskUpdatePriority[keyof typeof TaskUpdatePriority];


export const TaskUpdatePriority = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
} as const;

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  TODO: 'TODO',
  IN_PROGRESS: 'IN_PROGRESS',
  DONE: 'DONE',
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  priority?: TaskUpdatePriority;
  status?: TaskUpdateStatus;
  deadline?: string;
  /** @nullable */
  classificationId?: number | null;
  /** @nullable */
  groupId?: number | null;
  /** @nullable */
  assignedUserId?: number | null;
}

export interface Subtask {
  id: number;
  taskId: number;
  title: string;
  completed: boolean;
  createdAt: string;
}

export interface SubtaskInput {
  /** @minLength 1 */
  title: string;
}

export interface SubtaskUpdate {
  /** @minLength 1 */
  title?: string;
  completed?: boolean;
}

export type ClassificationType = typeof ClassificationType[keyof typeof ClassificationType];


export const ClassificationType = {
  DEFAULT: 'DEFAULT',
  CUSTOM: 'CUSTOM',
} as const;

export interface Classification {
  id: number;
  userId: number;
  name: string;
  type: ClassificationType;
  createdAt: string;
}

export interface ClassificationInput {
  /** @minLength 1 */
  name: string;
}

export interface DashboardSummary {
  totalTasks: number;
  todoCount: number;
  inProgressCount: number;
  doneCount: number;
  overdueCount: number;
  todayCount: number;
}

export interface GroupMember {
  id: number;
  userId: number;
  name: string;
}

export type GroupTaskStatus = typeof GroupTaskStatus[keyof typeof GroupTaskStatus];


export const GroupTaskStatus = {
  TODO: 'TODO',
  IN_PROGRESS: 'IN_PROGRESS',
  DONE: 'DONE',
} as const;

export type GroupTaskPriority = typeof GroupTaskPriority[keyof typeof GroupTaskPriority];


export const GroupTaskPriority = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
} as const;

export interface GroupTask {
  id: number;
  title: string;
  status: GroupTaskStatus;
  priority: GroupTaskPriority;
  assigneeName?: string | null;
  /** @nullable */
  assignedUserId?: number | null;
  subtaskCount: number;
  completedSubtaskCount: number;
}

export interface GroupNote {
  id: number;
  authorId: number;
  authorName: string;
  noteText: string;
  createdAt: string;
}

export interface Group {
  id: number;
  name: string;
  color?: string | null;
  createdBy: number;
  members: GroupMember[];
  tasks: GroupTask[];
  notes: GroupNote[];
}

export interface GroupInput {
  /** @minLength 1 */
  name: string;
  color?: string;
}

export interface UpdateGroupInput {
  /** @minLength 1 */
  name?: string;
  color?: string;
}

export interface GroupNoteInput {
  /** @minLength 1 */
  noteText: string;
  authorId?: number;
}

export interface AddGroupMemberInput {
  /** @minLength 1 */
  memberName: string;
}