# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Task Force API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups:
    get:
      operationId: getGroups
      tags: [groups]
      summary: List groups the current user belongs to or created
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a new group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{id}:
    patch:
      operationId: updateGroup
      tags: [groups]
      summary: Update group name and/or color (creator only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroupInput"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "403":
          description: Not the group creator
        "404":
          description: Group not found
    delete:
      operationId: deleteGroup
      tags: [groups]
      summary: Delete a group (creator only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Group deleted
        "403":
          description: Not the group creator
        "404":
          description: Group not found
  /groups/{id}/members:
    post:
      operationId: addGroupMember
      tags: [groups]
      summary: Add a member to a group by username
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddGroupMemberInput"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "404":
          description: Group or user not found
        "409":
          description: User already a member
  /groups/{id}/members/{memberId}:
    delete:
      operationId: removeGroupMember
      tags: [groups]
      summary: Remove a member from a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Member removed
        "403":
          description: Not authorized
        "404":
          description: Group or member not found
  /groups/{id}/notes:
    get:
      operationId: getGroupNotes
      tags: [groups]
      summary: List notes for a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupNote"
    post:
      operationId: createGroupNote
      tags: [groups]
      summary: Post a note to a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupNoteInput"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupNote"
  /groups/{id}/notes/{noteId}:
    delete:
      operationId: deleteGroupNote
      tags: [groups]
      summary: Delete a group note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Note deleted
        "403":
          description: Not authorized
        "404":
          description: Note not found
components:
  schemas:
...
    GroupMember:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
      required:
        - id
        - userId
        - name
    GroupTask:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        status:
          type: string
          enum: [TODO, IN_PROGRESS, DONE]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        assigneeName:
          type: string
          nullable: true
        assignedUserId:
          type: ["integer", "null"]
        subtaskCount:
          type: integer
        completedSubtaskCount:
          type: integer
      required:
        - id
        - title
        - status
        - priority
        - subtaskCount
        - completedSubtaskCount
    GroupNote:
      type: object
      properties:
        id:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        noteText:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - authorId
        - authorName
        - noteText
        - createdAt
    Group:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
          nullable: true
        createdBy:
          type: integer
        members:
          type: array
          items:
            $ref: "#/components/schemas/GroupMember"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/GroupTask"
        notes:
          type: array
          items:
            $ref: "#/components/schemas/GroupNote"
      required:
        - id
        - name
        - createdBy
        - members
        - tasks
        - notes
    GroupInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
      required:
        - name
    UpdateGroupInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
    GroupNoteInput:
      type: object
      properties:
        noteText:
          type: string
          minLength: 1
        authorId:
          type: integer
      required:
        - noteText
    AddGroupMemberInput:
      type: object
      properties:
        memberName:
          type: string
          minLength: 1
      required:
        - memberName