# lib/api-zod/src/generated/types/groupTask.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Task Force API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GroupTaskPriority } from './groupTaskPriority';
import type { GroupTaskStatus } from './groupTaskStatus';

export interface GroupTask {
  id: number;
  title: string;
  status: GroupTaskStatus;
  priority: GroupTaskPriority;
  assigneeName?: string | null;
  /** @nullable */
  assignedUserId?: number | null;
  subtaskCount: number;
  completedSubtaskCount: number;
}